requirements.txt
# Core Libraries
langchain==0.1.20
openai==1.35.1
google-api-python-client==2.146.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
oauth2client==4.1.3
pydantic==2.9.2
python-dotenv==1.0.1

# PDF Processing
pdfplumber==0.10.3

# Gmail API & MIME Handling
email-validator==2.1.0
beautifulsoup4==4.12.3

# Utilities
requests==2.32.3